Add files via upload
This was a little programming assignment working with logical programming and using the Prolog language. The assignment was to create 3 functions: a motherInLaw(X,Y) function that determines if X is Y's mother-in-law, a grandParent(X,Y) function which determines if X is Y's grandparent, and a grandMother(X,Y) function which determines if X is Y's grandmother.